init & delete sensor_msgs

+++ applanix_driver/src/applanix_driver_ros/conversions.cpp
@@ -338,9 +338,12 @@ autoware_sensing_msgs::msg::GnssInsOrientationStamped toAutowareOrientationMsg(
     autoware_sensing_msgs::msg::GnssInsOrientationStamped autowareOrientationMsg =
             toAutowareOrientationMsg(ins_solution, enable_ned2enu_transform);
 
-    autowareOrientationMsg.orientation.rmse_rotation_x = covariance.attitude_rms.roll / 180 * M_PI;
-    autowareOrientationMsg.orientation.rmse_rotation_y = covariance.attitude_rms.pitch / 180 * M_PI;
-    autowareOrientationMsg.orientation.rmse_rotation_z = covariance.attitude_rms.heading / 180 * M_PI;
+    autowareOrientationMsg.orientation.rmse_rotation_x =
+            covariance.attitude_rms.roll / 180.0f * static_cast<float>(M_PI);
+    autowareOrientationMsg.orientation.rmse_rotation_y =
+            covariance.attitude_rms.pitch / 180.0f * static_cast<float>(M_PI);
+    autowareOrientationMsg.orientation.rmse_rotation_z =
+            covariance.attitude_rms.heading / 180.0f * static_cast<float>(M_PI);
 
     return autowareOrientationMsg;
 }
